Major cleanup pass for bevy_egui
