+++ crates/bevy_egui/src/egui_node.rs
@@ -1,20 +1,32 @@
-use std::{borrow::Cow, collections::HashMap, collections::HashSet, sync::Arc};
-
-use bevy::{prelude::*, render::mesh::VertexAttribute, render::mesh::VertexAttributeValues, render::pipeline::DynamicBinding, render::pipeline::PipelineSpecialization, render::pipeline::PrimitiveTopology, render::pipeline::RenderPipeline, render::render_graph::ResourceSlots, render::renderer::RenderContext, render::renderer::RenderResourceContext, render::texture::TextureFormat};
-use bevy::render::{
-    renderer::RenderResources,
-    shader::ShaderDefs,
-    render_graph::{
-        SystemNode,
-        Node,
-        CommandQueue
-    },
+use bevy::{
+    prelude::*,
+    render::{
+        shader::ShaderDefs,
+        texture::TextureFormat,
+        render_graph::{
+            Node,
+            SystemNode,
+            CommandQueue,
+            ResourceSlots,
+        },
+        renderer::{
+            RenderContext,
+            RenderResourceContext,
+            RenderResources
+        }
+    }
 };
-use egui::{RawInput, Srgba, Rgba};
 
-use crate::{egui_ui::EguiFrameStartEvent, mesh_handler::MeshHandler};
+use crate::{
+    egui_ui::EguiFrameStartEvent,
+    mesh_handler::MeshHandler,
+    components::EguiJobsDescriptor,
+    systems::EguiContext
+};
 
-const FAKE_RAW_INPUT: egui::RawInput = egui::RawInput {
+// TODO(#60) actually give real input to egui
+/// Default fake input so the gui can at least work.
+pub const FAKE_RAW_INPUT: egui::RawInput = egui::RawInput {
     mouse_down: false,
     mouse_pos: None,
     scroll_delta: egui::math::vec2(0.0, 0.0),
@@ -24,8 +36,8 @@ const FAKE_RAW_INPUT: egui::RawInput = egui::RawInput {
     events: Vec::new()
 };
 
-// TODO: change this to use an Option<Handle<Texture>> with a shader def so we don't have this weird default crap
-// TODO: finally split into separate components for the egui rendering passes and the egui render resources
+// TODO(#57): finally split into separate components for the egui rendering passes and the egui render resources
+// TODO(#58) change this to use an Option<Handle<Texture>> with a shader def so we don't have this weird default crap
 /// Represents the egui rendering passes and systems.
 /// Entities with this component are expected to represent rendering to the specified context.
 /// This node, with the context name intended for entities to operate on, is also added as a system node.
@@ -59,107 +71,12 @@ impl EguiNode {
     }
 }
 
-use std::ops::Range;
-
-#[derive(Default)]
-pub struct EguiJobsDescriptor {
-    pub jobs: Vec<(Range<u32>, i32)>
-}
-
-pub struct EguiId(&'static str);
-
-impl Default for EguiId {
-    fn default() -> Self {
-        Self("DEFAULT")
-    }
-}
-
-#[derive(Bundle)]
-pub struct EguiComponents {
-
-    pub id: EguiId,
-    pub jobs_descriptor: EguiJobsDescriptor,
-    /// Associates our entity with the Egui render passes and holds information to associate with an Egui context.
-    pub egui_node: Handle<EguiNode>,
-    pub mesh: Handle<Mesh>,
-    pub render_pipelines: RenderPipelines,
-    pub draw: Draw,
-    pub transform: Transform,
-    pub global_transform: GlobalTransform,
-}
-
-impl Default for EguiComponents {
-    fn default() -> Self {
-        Self {
-            id: Default::default(),
-            jobs_descriptor: Default::default(),
-            egui_node: Default::default(),
-            mesh: Default::default(),
-            render_pipelines: RenderPipelines::from_pipelines(vec![RenderPipeline::specialized(
-                crate::base_setup::EGUI_PIPELINE_HANDLE,
-                PipelineSpecialization {
-                    dynamic_bindings: vec![
-                        // Transform
-                        DynamicBinding {
-                            bind_group: 2,
-                            binding: 0
-                        },
-                    ],
-                    ..Default::default()
-                },
-            )]),
-            draw: Default::default(),
-            transform: Default::default(),
-            global_transform: Default::default(),
-        }
-    }
-}
-
-impl EguiComponents {
-    pub(crate) fn from_setup(resources: &Resources) -> Self {
-        let mut meshes = resources.get_mut::<Assets<Mesh>>().unwrap();
-
-        Self {
-            // mesh: meshes.add(Mesh::from(shape::Cube::default())),
-            mesh: meshes.add(Mesh {
-                primitive_topology: PrimitiveTopology::TriangleList,
-                attributes: vec![
-                    VertexAttribute {
-                        name: "BevyEguiVertex_Position".into(),
-                        values: VertexAttributeValues::Float3(Vec::new())
-                    },
-                    VertexAttribute {
-                        name: "BevyEguiVertex_Normal".into(),
-                        values: VertexAttributeValues::Float3(Vec::new()),
-                    },
-                    VertexAttribute {
-                        name: "BevyEguiVertex_Uv".into(),
-                        values: VertexAttributeValues::Float2(Vec::new()),
-                    },
-                    VertexAttribute {
-                        name: "BevyEguiVertex_Color".into(),
-                        values: VertexAttributeValues::Float4(Vec::new())
-                    },
-                    VertexAttribute {
-                        name: "BevyEguiVertex_ClipMin".into(),
-                        values: VertexAttributeValues::Float2(Vec::new())
-                    },
-                    VertexAttribute {
-                        name: "BevyEguiVertex_ClipMax".into(),
-                        values: VertexAttributeValues::Float2(Vec::new())
-                    }
-                ],
-                indices: Some(Vec::new()),
-            }),
-            ..Default::default()
-        }
-    }
-}
-
-#[derive(Clone, Default)]
+/// Render system node to do render utility updates for a given egui context
+#[derive(Clone)]
 pub struct EguiSystemNode {
     /// The egui context this is attached to
-    // pub context: Arc<egui::Context>,
+    pub context: Handle<EguiContext>,
+    /// Command queue to assist with render resource operations
     pub command_queue: CommandQueue,
 }
 
@@ -182,33 +99,20 @@ impl SystemNode for EguiSystemNode {
         println!("Grabbing EguiSystemNode system!");
         let system = egui_node_system.system();
 
-        let mut context = egui::Context::new();
-
-        // TODO: provide real RawInput so things aren't possibly bad!
-        let _ = context.begin_frame(FAKE_RAW_INPUT);
-
         commands.insert_local_resource(
             system.id(),
-            EguiSystemNodeState {
-                context,
-                draw_entity: None,
-
-                command_queue: self.command_queue.clone(),
-
-                mesh: None,
-                mesh_handler: MeshHandler::default(),
-                egui_node: None,
-                // texture_handler: TextureHandler::default(),
-            }
+            EguiSystemNodeState::from_node(self)
         );
 
         system
     }
 }
 
+// TODO(#59) modify EguiSystemNodeState so that it only holds onto the draw entity and gets the other information each tick from the query
+/// Internal state for the egui system node update system
 struct EguiSystemNodeState {
     /// The egui context this system node is associated with.
-    context: Arc<egui::Context>,
+    context: Handle<EguiContext>,
     /// The entity associated with the egui context which actually enables drawing of the ui.
     draw_entity: Option<Entity>,
 
@@ -222,8 +126,19 @@ struct EguiSystemNodeState {
 
     /// Handle to the egui node which stores texture data of this egui.
     egui_node: Option<Handle<EguiNode>>,
-    // The Handler which processes updates to the texture of font data.
-    // texture_handler: TextureHandler,
+}
+
+impl EguiSystemNodeState {
+    fn from_node(node: &EguiSystemNode) -> Self {
+        Self {
+            context: node.context,
+            draw_entity: None,
+            command_queue: node.command_queue.clone(),
+            mesh: None,
+            mesh_handler: MeshHandler::default(),
+            egui_node: None,
+        }
+    }
 }
 
 impl FromResources for EguiSystemNodeState {
@@ -232,15 +147,17 @@ impl FromResources for EguiSystemNodeState {
     }
 }
 
-// TODO: modify EguiSystemNodeState so that it only holds onto the draw entity and gets the other information each tick from the query
+/// Update system for a egui system node
+#[allow(clippy::too_many_arguments)]
+#[allow(clippy::type_complexity)]
 fn egui_node_system(
     mut state: Local<EguiSystemNodeState>,
-    mut frame_init_event: ResMut<Events<crate::EguiFrameStartEvent>>,
+    mut frame_init_event: ResMut<Events<EguiFrameStartEvent>>,
 
     mut nodes: ResMut<Assets<EguiNode>>,
+    mut contexts: ResMut<Assets<EguiContext>>,
 
     render_resource_context: Res<Box<dyn RenderResourceContext>>,
-    mut meshes: ResMut<Assets<Mesh>>,
     mut textures: ResMut<Assets<Texture>>,
 
     mut query: Query<(Entity, &mut EguiJobsDescriptor, &mut Handle<EguiNode>, &mut Handle<Mesh>, &mut RenderPipelines)>,
@@ -260,6 +177,8 @@ fn egui_node_system(
     } = &mut *state;
 
     let render_resource_context = &**render_resource_context;
+
+    let EguiContext { context, .. } = contexts.get_mut(context).unwrap();
 
     // If we've not set the draw entity yet, check if there exists the entity.
     // If it doesn't exist, we return and check again next tick.
@@ -314,18 +233,17 @@ fn egui_node_system(
 
     let mut entity_query = query.entity(*draw_entity)
         .expect("Unable to make query from egui draw entity!");
-    let (_, mut jobs_descriptor, node, _, pipelines) = entity_query.get()
+    let (_, mut jobs_descriptor, _, _, pipelines) = entity_query.get()
         .expect("Egui draw entity did not have a RenderPipelines or EguiJobsDescriptor for some reason?");
 
-    // TODO: update mesh handler to use staging buffers via the command_queue
+    // TODO(#54): update mesh handler to use staging buffers via command_queue to make this less bad
     mesh_handler.with_context(render_resource_context, *mesh)
         .update_from_jobs(jobs, &mut jobs_descriptor)
         .set_pipeline_bindings(pipelines);
 
     println!("Trying to send EguiFrameStart event!");
 
-    // TODO: eventually we'll need to give proper input _not_ at the end of the last frame.
-    // TODO: to accomplish this, the context should actually be owned by the frame start system and sent here.
+    // TODO(#60) actually give real input to egui, should require the context to actually be owned by the state update system and sent here rather than the other way around
     // Now that frame update crap is over, begin a new frame and send the Ui object away to be used
     frame_init_event.send(EguiFrameStartEvent {
         new_ui: context.begin_frame(FAKE_RAW_INPUT)
